+++ CHANGELOG.md
@@ -80,5 +80,7 @@
 - ~~ensure all features are documented~~
 - ~~document limitations when service provier only allows 1 active token per client_id~~
 - ~~activity based session expiry (floating window sessions)~~
-- jwt assertions
+- ~~jwt assertions~~
 - ~~forward auth plugin~~
+- ~~assertion query engines~~
+- ~~pipeline circuit breakers (`pcb`)~~